+++ src/stactools/noaa_cdr/sea_ice_concentration/constants.py
@@ -27,10 +27,10 @@ DESCRIPTION = (
     "ice, and monitor changes in sea ice concentration. The CDR "
     "combines concentration estimates using two algorithms "
     "developed at the NASA Goddard Space Flight Center (GSFC). "
-    "Gridded brightness temperatures acquired from a number of "
-    "Defense Meteorological Satellite Program (DMSP) passive "
-    "microwave radiometers provide the necessary input to "
-    "produce the dataset."
+    "Gridded brightness temperatures acquired from Defense "
+    "Meteorological Satellite Program (DMSP) passive microwave "
+    "radiometers, and, beginning 1 January 2025, the JAXA AMSR2 "
+    "sensor, provide the necessary input to produce the dataset."
 )
 EXTENT = Extent(
     spatial=SpatialExtent(
@@ -38,13 +38,13 @@ EXTENT = Extent(
     ),
     temporal=TemporalExtent([[datetime.datetime(1978, 10, 25, tzinfo=tzutc()), None]]),
 )
-DOI = "10.7265/efmz-2t65"
+DOI = "10.7265/b18j-z797"
 CITATION = (
-    "Meier, W. N., F. Fetterer, A. K. Windnagel, "
-    "and S. Stewart. 2021. NOAA/NSIDC Climate Data Record of "
-    "Passive Microwave Sea Ice Concentration, Version 4. "
+    "Meier, W. N., F. Fetterer, A. K. Windnagel, J. S. Stewart, "
+    "and T. Stafford. 2026. NOAA/NSIDC Climate Data Record of "
+    "Passive Microwave Sea Ice Concentration, Version 6. "
     "[Indicate subset used]. Boulder, Colorado USA. NSIDC: National "
-    "Snow and Ice Data Center https://doi.org/10.7265/efmz-2t65. "
+    "Snow and Ice Data Center https://doi.org/10.7265/b18j-z797. "
 )
 PROVIDERS = [
     Provider(
@@ -64,7 +64,7 @@ PROVIDERS = [
             ProviderRole.LICENSOR,
             ProviderRole.HOST,
         ],
-        "https://nsidc.org/data/g02202/versions/4",
+        "https://nsidc.org/data/g02202/versions/6",
     )
 ]
 KEYWORDS = COMMON_KEYWORDS + [
@@ -88,7 +88,7 @@ HOMEPAGE_LINK = Link(
 )
 # Want to include this because the user guide will always be findable here.
 DATA_HOMEPAGE_LINK = Link(
-    target=("https://nsidc.org/data/g02202/versions/4"),
+    target=("https://nsidc.org/data/g02202/versions/6"),
     rel="about",
     media_type=MediaType.HTML,
     title="CDR Homepage",