+++ tests/sea_ice_concentration/test_commands.py
@@ -6,7 +6,7 @@ from .. import run_command, test_data
 
 
 def test_create_item(tmp_path: Path) -> None:
-    path = test_data.get_path("data-files/seaice_conc_daily_nh_20211231_f17_v04r00.nc")
+    path = test_data.get_path("data-files/sic_psn25_19850130_n07_v06r00.nc")
     result = run_command(
         f"noaa-cdr sea-ice-concentration create-item {path} {tmp_path}/out.json"
     )
@@ -16,7 +16,7 @@ def test_create_item(tmp_path: Path) -> None:
 
 
 def test_create_item_with_cogs(tmp_path: Path) -> None:
-    path = test_data.get_path("data-files/seaice_conc_daily_nh_20211231_f17_v04r00.nc")
+    path = test_data.get_path("data-files/sic_psn25_19850130_n07_v06r00.nc")
     result = run_command(
         f"noaa-cdr sea-ice-concentration create-item --cogs {path} {tmp_path}/out.json"
     )